Chapter 4: Expand process test coverage

+++ src/process.cpp
@@ -25,22 +25,22 @@ sdb::stop_reason::stop_reason(int wait_status)
     if (WIFEXITED(wait_status))
     {
         reason = process_state::Exited;
-        info = WEXITSTATUS(wait_status); 
+        info = WEXITSTATUS(wait_status);
     }
     else if (WIFSIGNALED(wait_status))
     {
         reason = process_state::Terminated;
-        info = WTERMSIG(wait_status); 
+        info = WTERMSIG(wait_status);
     }
     else if (WIFSTOPPED(wait_status))
     {
         reason = process_state::Stopped;
-        info = WSTOPSIG(wait_status); 
+        info = WSTOPSIG(wait_status);
     }
 
 }
 
-std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path)
+std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path, bool attach)
 {
     pipe channel(true); // We have to call pipe before we call fork()
     pid_t pid;
@@ -53,7 +53,7 @@ std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path)
     {
         // We are in the child process
         channel.close_read();
-        if (ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0)
+        if (attach && ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0)
         {
             exit_with_perror(channel, "Tracing failed");
         }
@@ -79,8 +79,13 @@ std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path)
         }
     }
 
-    std::unique_ptr<process> proc{new process{pid, true}}; 
-    proc->wait_on_signal();
+    std::unique_ptr<process> proc{new process{pid, true, attach}};
+
+    if (attach)
+    {
+        proc->wait_on_signal();
+    }
+
     return proc;
 }
 
@@ -96,7 +101,7 @@ std::unique_ptr<sdb::process> sdb::process::attach(pid_t pid)
         error::send_errno(std::format("Could not attach to pid {}", pid));
     }
 
-    std::unique_ptr<process> proc{new process{pid, false}}; 
+    std::unique_ptr<process> proc{new process{pid, false, true}};
     proc->wait_on_signal();
 
     return proc;
@@ -107,14 +112,17 @@ sdb::process::~process()
     if (pid_ != 0)
     {
         int status;
-        if (state_ == process_state::Running)
+        if (is_attached_)
         {
-            kill(pid_, SIGSTOP);
-            waitpid(pid_, &status, 0);
-        }
+            if (state_ == process_state::Running)
+            {
+                kill(pid_, SIGSTOP);
+                waitpid(pid_, &status, 0);
+            }
 
-        ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
-        kill(pid_, SIGCONT);
+            ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
+            kill(pid_, SIGCONT);
+        }
 
         if (terminate_on_end_)
         {
@@ -128,7 +136,7 @@ void sdb::process::resume()
 {
     if (ptrace(PTRACE_CONT, pid_, nullptr, nullptr) < 0)
     {
-        error::send_errno("Could not resume"); 
+        error::send_errno("Could not resume");
     }
 
     state_ = process_state::Running;
@@ -140,7 +148,7 @@ sdb::stop_reason sdb::process::wait_on_signal()
     int options = 0;
     if (waitpid(pid_, &wait_status, options) < 0)
     {
-        error::send_errno("waitpid failed"); 
+        error::send_errno("waitpid failed");
     }
 
     stop_reason reason(wait_status);
